(103-153)
Collects the National ID number (NIDnumber) for each voter.
Checks if the National ID number is less than or equal to 50, and if so, allows the voter to proceed with voting.Displays a list of political parties and prompts the voter to select one.Validates the vote choice, casting the vote using the castVote function.
Updates the linked list of voters (head and current) using the createVoter function.If the National ID number is greater than 50, the voter is considered ineligible to vote.